Add duplicate-removal functions
"When in doubt, throw it out" - wrong data cannot be tolerated.  A common error that can be found in datasets is duplicate rows.  This is when one row claims that, at location x on day y, z mosquitoes were collected, while another row claims something else of locations x on day y.  Since both rows cannot be correct at the same time, both rows must be removed.  Of course, seeing such errored data may raise doubts about the collection it came from as a whole.  Therefore, as it is necessary to throw out data when in doubt of its accuracy, any collection containing a duplicate row must be thrown out altogether.  This checking for doubt and throwing out can be done quickly and painlessly with the `removeDuplicates(df)` function.  This function will remove all collections containing duplicates and return the new dataframe.  If, however, it is only necessary to throw out duplicate rows and not entire collections, the `removeDuplicateRows(df)` function can be used instead.  Either function will print a message to the console explaining what amount of data was removed.